pyproject.toml
[project]
name = "agentic-test"
version = "0.1.0"
description = "MCP-oriented testing demo with a MongoDB-backed microblog REST API (FastAPI)"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pymongo>=4.6.0",
    "pydantic-settings>=2.0.0",
    "email-validator>=2.0.0",
    "mcp>=1.27.0",
    "playwright>=1.58.0",
    "httpx>=0.28.1",
]

[project.scripts]
agentic-test = "agentic_test.main:main"
microblog-api = "agentic_test.api.cli:main"
agentic-mcp-server = "agentic_test.ai_api_tester.server:main"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm.build]
package-dir = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "generated_api: Playwright APIRequestContext tests produced by the AI harness",
    "handwritten_api: human-maintained API tests under tests/handwritten_api",
    "integration: tests that need Playwright browsers or longer I/O",
]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
namespace_packages = true
explicit_package_bases = true
mypy_path = "src"
packages = ["agentic_test.ai_api_tester"]

[tool.coverage.run]
source_pkgs = ["agentic_test"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = ["pragma: no cover", "if TYPE_CHECKING:"]
fail_under = 0

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
    "mongomock>=4.1.0",
    "pyyaml>=6.0",
    "ruff>=0.15.9",
    "mypy>=1.20.0",
    "pytest-cov>=7.1.0",
    "pytest-json-report>=1.5.0",
    "pytest-httpserver>=1.1.5",
    "types-PyYAML>=6.0.12.20250915",
]
